+++ posts_app/views/delete_post.py
@@ -9,6 +9,8 @@ from posts_app.models import Post
 
 class DeletePpost(View):
     def get(self, request, pk):
-        get_post = Post.objects.get(pk=pk)
-        get_post.delete()
-        return redirect(reverse('profile'))
+        if request.user.is_authenticated:
+            get_post = Post.objects.get(pk=pk)
+            get_post.delete()
+            return redirect(reverse('profile'))
+        return redirect("login")
